Move Rating and ModeratorStatement Factories up
