+++ tests/exports/conftest.py
@@ -4,6 +4,4 @@ from tests.exports import factories
 from tests.ideas import factories as ideas_factories
 
 register(ideas_factories.IdeaFactory)
-register(factories.RatingFactory)
-register(factories.ModeratorStatementFactory)
 register(factories.ProposalFactory)
